Post out to Twitter
Our canonical parameter will be `body`, and for WordPress or Tumblr, also
`title`.

If Twitter responds with an error, we parse it and hoist it into
`response.twitter.error` so the developer can handle things consistently. We may
want to also return the original error, but I'll leave that until there's a use
case.

Some errors don't make sense or can be actively confusing, like "Client must
provide a 'status' parameter with a value.", and we'll have to handle those
intelligently for the developer. The fix for that one is coming up next.

Closes #482