Post statuses to Yammer
Yammer's API is a little more finicky than some. If you don't include an access
token, they just return `' '` instead of something useful, and I imagine they'll
do that in other cases as well. For now, it's just caught with a generic "Yammer
didn't return JSON.", and as we encounter more, we can figure out what to do.

Closes #495