Post statuses to LinkedIn
Though LinkedIn's
[share api](https://developer.linkedin.com/documents/share-api) says you should
pass it a top-level `share` object, that only applies to XML. When using JSON,
they appear to wrap the whole POST body in the `share`.

As noted in the comments, success gives you back an empty HTTP 201 with a
`Location` header telling you where the share is. Their docs note that the URL
provided there isn't actually available, but it's a source of the share ID that
we can use in our own success response. We can parse the ID out if we like, but
it'll be irrelevant if we return the full share data after #498.

Closes #494