Include relevant access tokens in app info
This will be handy when we want to tie our original list of tokens to the apps
we got out. We could return the original hash of `token: info`, but that's
highly redundant. It might be useful to make that an option though.

I added `tokenCount` for ease when you're looking in a browser.