Post photos to Tumblr
The Tumblr API is crazy and this is the only library I can find that manages to
upload a photo. It does it with a bunch of crazy regex substitutions during
OAuth, which are frightening but appear to work.

Something really weird is still happening, and we can't send captions along with
photos, so for now I just ignore any body included with the upload. I also don't
know what effect trying to include Tumblr-specific parameters might have.