LinkedIn connections wait for network updates.
LinkedIn now only runs connections if there has been a PROF or NCON network
update
(https://developer.linkedin.com/documents/get-network-updates-and-statistics-api).

Also, the fresh=true param, now looks for a synclets field called `freshDeps` to
ensure that all synclets that are required to ensure freshness get run (and in
the right order). In this case, connections depends on network.