a migration to fix corrupted github auth objects.
early on, we were saving events data to the auth object (wtf?) which very often
causes it to be more tha 65k chars and would get cut off when being written to
MySQL. This meant that the next time it was retrieved it could no longer be
parsed. Luckily, all of the important auth info (tokens, apps, accounts) are
right at beginning, so they didn't get lost.

This just trims the text before the events info, closes the JSON object,
reparses to make it it's valid, then checks for all the imporant fields and
updates the DB row.