Accept multiple `to` or `from` parameters
If your request has multiple `to` parameters, eg

    POST /types/statuses?to=foo&to=bar

the parameter comes in as

    ['foo', 'bar']

which wouldn't allow `list.match` in `parseProfileList`, which was expecting a
string.

This just makes sure it's a string before it makes sure it's an array, which is
sort of sad but will be most robust to people including multiple,
comma-separated `to` parameters.