Accept service params that are already JSON
In the first implementation, I assumed that `*_params` would be a string, but
sometimes (like if you use `facebook_params[foo]=bar` syntax), Express can parse
it into an object. This accepts that.

We were also making the request anyway (potentially without the service
parameters), but you'd see an error in the response that we couldn't parse it.
Now only see that error if we definitely didn't try a request.