Make linting ignore custom Zeo OAuth module
This module is the source of thousands of linting errors, and in particular some
of them are `==` vs `===` differences that might actually be relied on by the
code. Ultimately, it should be moved to a `node_module`, and ideally we should
figure out exactly the difference and see if there's a better solution than
using a custom OAuth. But since it was introduced in one commit (6b10c35),
that's going to be tricky.

In the meantime, let's ignore it during lint and focus on our own code.