Fix behavior after scope fix
The code in `lib/friends#bump` was actually relying on mismanaged variable
scope, and when that got fixed, `indexMe` started crashing. (In fact, indexMe
_always_ crashes the way we're calling it, but because of the scope issue, it
was never actually running).

Here's what happened:
- `var self` was declared in `bump` and set to `false`.
- Immediately, `cset.forEach` opened a new scope. Halfway through, `var self`
  was declared again.
- This created a new `self` that was hoisted inside `forEach`'s function and
  manipulated
- When the `forEach` finished, the inner `self` was thrown away (its scope
  closed), and we came back to the outer `self`, which was still `false`.
- The outer `self`  was passed to `indexMe` as `entry`, and `indexMe` promptly
  returned, accomplishing nothing.

When I cleaned up the doubly-declared `self`, suddenly we began manipulating the
outer scope's `self` and ended up passing a profile ID (eg `'34301'`) into
`indexMe`. `indexMe` tried to parse it as an IDR, but it isn't, and the worker
or dawg would crash out.

To fix this, I removed the outer `self` and passed `false` to `indexMe`. This
looks all sorts of daft, but it's what the code was doing before and I bet
someone who knows what's supposed to be going on will want to fix it. In the
meantime, `indexMe` will continue doing what it's doing on `master`: zilch.

cc @quartzjer