Show a real stack trace in `logger.error`
If `logger.error` receives an `Error` object, it will do the helpful thing and
print the error's real stacktrace. If it receives a string, it will make up the
same trace it's been using so far, which shows where the string was produced.

This makes it way easier to track down uncaught exceptions in particular, and
paves the transition towards all of our string errors becoming real objects.